Complete steps to set up for VATT inference
Follow the steps contained in this script to set up the environment for VATT inference. VATT-LLama-T are used as an example in the script, and other models should follow similar procedures. Here we assume that the eva-CLIP visual features are already extracted, and demonstrate how we generate the audio on VGGSound dataset.